asmr

+++ C++/DortyLibs/BigInt/core.cpp
@@ -72,8 +72,8 @@ bool BigInt::even() const {
 /// [5,4,3,2,1](12345) (length=3)-> [0,0,0,5,4,3,2,1](12345000)
 /// Следует Улучшить. Ужасно плохо реализованно
 void inline BigInt::_appendZeros(int length){
-    std::vector<int> v1(length);
-    std::vector<int> tmp = _digits;
+    std::vector<long long> v1(length);
+    std::vector<long long> tmp = _digits;
     _digits.clear();
     std::merge(v1.begin(), v1.end(), tmp.begin(), tmp.end(), std::back_inserter(_digits));
 }
